+++ app/build.gradle
@@ -41,6 +41,7 @@ android {
 }
 
 dependencies {
+    def roomVersion = "2.2.6"
     implementation 'androidx.core:core-ktx:1.7.0'
     implementation 'androidx.appcompat:appcompat:1.4.0'
     implementation 'com.google.android.material:material:1.4.0'
@@ -57,6 +58,10 @@ dependencies {
     implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-core:1.5.2'
     implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-android:1.5.2'
     implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+
+    //Room
+    implementation "androidx.room:room-ktx:$roomVersion"
+    kapt "androidx.room:room-compiler:$roomVersion"
     
     //LifeCycle
     implementation 'androidx.lifecycle:lifecycle-common:2.4.0'
@@ -79,11 +84,6 @@ dependencies {
     implementation 'com.squareup.picasso:picasso:2.71828'
     implementation 'androidx.recyclerview:recyclerview:1.2.1'
 
-    //Testing Coroutine
-    testImplementation "org.jetbrains.kotlinx:kotlinx-coroutines-test:1.5.0"
-
-    //Thuth
-    testImplementation "com.google.truth:truth:1.1.3"
     //MockWebServer
     testImplementation "com.squareup.okhttp3:mockwebserver:4.9.3"
 
@@ -100,4 +100,29 @@ dependencies {
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
     // Testing Navigation
     androidTestImplementation("androidx.navigation:navigation-testing:2.3.5")
+
+    // Local unit tests
+    testImplementation "androidx.test:core:1.4.0"
+    testImplementation "junit:junit:4.13.2"
+    testImplementation "androidx.arch.core:core-testing:2.1.0"
+    testImplementation "org.jetbrains.kotlinx:kotlinx-coroutines-test:1.5.1"
+    testImplementation "com.google.truth:truth:1.1.3"
+    testImplementation "com.squareup.okhttp3:mockwebserver:4.9.1"
+    testImplementation 'com.nhaarman.mockitokotlin2:mockito-kotlin:2.2.0'
+    testImplementation 'org.mockito:mockito-inline:2.21.0'
+    debugImplementation "androidx.compose.ui:ui-test-manifest:1.1.0-alpha04"
+
+    // Instrumentation tests
+    androidTestImplementation 'com.google.dagger:hilt-android-testing:2.37'
+    kaptAndroidTest 'com.google.dagger:hilt-android-compiler:2.37'
+    androidTestImplementation "junit:junit:4.13.2"
+    androidTestImplementation "org.jetbrains.kotlinx:kotlinx-coroutines-test:1.5.1"
+    androidTestImplementation "androidx.arch.core:core-testing:2.1.0"
+    androidTestImplementation "com.google.truth:truth:1.1.3"
+    androidTestImplementation 'androidx.test.ext:junit:1.1.3'
+    androidTestImplementation 'androidx.test:core-ktx:1.4.0'
+    androidTestImplementation "com.squareup.okhttp3:mockwebserver:4.9.1"
+    androidTestImplementation 'com.nhaarman.mockitokotlin2:mockito-kotlin:2.2.0'
+    androidTestImplementation 'org.mockito:mockito-inline:2.21.0'
+    androidTestImplementation 'androidx.test:runner:1.4.0'
 }
