+++ app/src/main/java/com/yuri/apprickmorty/domain/usescases/SavePersogensFavoritosUseCase.kt
@@ -1,0 +1,10 @@
+package com.yuri.apprickmorty.domain.usescases
+
+import com.yuri.apprickmorty.data.models.Personagem
+import com.yuri.apprickmorty.domain.repository.PersonagemRepository
+
+class SavePersogensFavoritosUseCase(
+    private val repository: PersonagemRepository
+) {
+    suspend fun execute(personagem: Personagem) = repository.savePersonagem(personagem)
+}
