+++ app/src/main/java/com/yuri/apprickmorty/ui/views/MainActivity.kt
@@ -1,0 +1,39 @@
+package com.yuri.apprickmorty.ui.views
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import androidx.navigation.NavController
+import androidx.navigation.fragment.NavHostFragment
+import androidx.navigation.ui.setupActionBarWithNavController
+import androidx.navigation.ui.setupWithNavController
+import com.yuri.apprickmorty.R
+import com.yuri.apprickmorty.databinding.ActivityMainBinding
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
+class MainActivity : AppCompatActivity() {
+
+    private var navController: NavController? = null
+    private lateinit var binding: ActivityMainBinding
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        binding = ActivityMainBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+
+        val navHostFragment = supportFragmentManager
+            .findFragmentById(R.id.newsFragment) as NavHostFragment
+
+        navController = navHostFragment.navController
+
+
+        setupActionBarWithNavController(navController!!)
+        binding.bnvPersonagens.setupWithNavController(
+            navController!!
+        )
+    }
+    override fun onSupportNavigateUp(): Boolean {
+        return navController!!.navigateUp() || super.onSupportNavigateUp()
+    }
+
+}
