+++ app/src/main/res/layout/fragment_detalhe_personagem.xml
@@ -10,9 +10,10 @@
         android:layout_height="match_parent"
         android:background="@drawable/background_pesquisa"
         android:elevation="0dp"
-        tools:context=".ui.main.views.DetalhePersonagemFragment">
+        tools:context=".ui.views.DetalhePersonagemFragment">
 
         <androidx.constraintlayout.widget.ConstraintLayout
+            android:id="@+id/constraintLayout"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
             android:layout_margin="@dimen/dimensao_32_dp"
@@ -188,6 +189,18 @@
                 app:layout_constraintStart_toStartOf="parent"
                 app:layout_constraintTop_toBottomOf="@+id/textViewPersonagemNomeDetalhes"
                 tools:text="HUMAN" />
+
+            <com.google.android.material.floatingactionbutton.FloatingActionButton
+                android:id="@+id/floatingActionButtonSavePersonagemFavorito"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_marginEnd="16dp"
+                android:layout_marginBottom="16dp"
+                android:clickable="true"
+                app:layout_constraintBottom_toBottomOf="parent"
+                app:layout_constraintEnd_toEndOf="parent"
+                app:srcCompat="@drawable/ic_baseline_save_24"
+                tools:ignore="SpeakableTextPresentCheck" />
 
         </androidx.constraintlayout.widget.ConstraintLayout>
