+++ app/src/main/res/layout/fragment_personagens_favoritos.xml
@@ -1,15 +1,18 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<layout
+    xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
+    >
+<androidx.constraintlayout.widget.ConstraintLayout
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/black_ligth"
     android:orientation="vertical"
-    tools:context=".ui.main.views.PersonagensFavoritosFragment">
+    tools:context=".ui.views.PersonagensFavoritosFragment">
 
     <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/recycclerviewPersonagens"
+        android:id="@+id/recycclerviewPersonagensFavoritos"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         app:layout_constraintBottom_toBottomOf="parent"
@@ -27,3 +30,4 @@
     </androidx.recyclerview.widget.RecyclerView>
 
 </androidx.constraintlayout.widget.ConstraintLayout>
+</layout>
